+++ src/main/java/com/example/emailapp/util/StringUtil.java
@@ -1,0 +1,10 @@
+package com.example.emailapp.util;
+
+public class StringUtil {
+    public StringUtil() {
+    }
+
+    public static boolean isValid(String data) {
+        return null != data && data.trim().length() > 0;
+    }
+}