+++ versionHistory.md
@@ -7,4 +7,10 @@
 2. V2.0 - 12 Mar 2021
     * ToDo.md added
     * versionHistory.md added
+   
+3. V3.0 - 12 Mar 2021, Friday
+
+   * SMTP config moved to `application.resources` properties file
+   * Added `EmailConfig.java`, `StringUtil.java` files
+